[minor] Expired sessions on startup are now considerd as not logged in.
When the Desktop started, it checked to see if there was a current user. If there was, regardless of whether the session id was still valid or not we were launching the Desktop with it. This meant that further down if the session was expired we were going to be asked for our credentials.

Now, if the session is expired, we clear the cache and ask for a user again. This will prompt for the full set of credentials and site.

Closes https://github.com/shotgunsoftware/tk-framework-desktopstartup/pull/21