update bundled tk-core to the latest v0.17.12 release
for #35517: new core version adds support for setting connection timeout when attempting to connect to the Toolkit app store. In cases where internet access is blocked, this ensures that SG Desktop will abort the auto-update of the startup framework within a reasonable amount of time and continue loading without error.