Commented out every setting by default.
Some people are copying this file straight into their user folder and then tweak some settings. However it is easy to miss a line that needs to be commented out and pick up a setting we don't want.

To fix this issue, everything is now commented out by default and you need to opt it instead of opting out.