For #46042: Restores the global debug logging setting prior to core swap. (#52)
This is resolving an issue that arises when the tk-core version in the site config is older than v0.18.117. In that case, when the global debug is toggled on in desktopstartup, TK_DEBUG is set in the environment. If it is toggled off after the core swap, that environment variable isn't purged since the older tk-core doesn't know to do that. We end up in a situation where TK_DEBUG persists, resulting in any subprocesses spawned running in debug logging active when it wasn't asked for.

Instead, we now restore the global debug setting prior to core swap, which ensures that the same version of tk-core is used both to set and restore global debug.